pm/soc/*/host_idx_mapping.c: Fix all MISRA.ETYPE.ASSIGN.2012 issues
Fixes: SYSFW-1510

+++ pm/soc/am6x/host_idx_mapping.c
@@ -89,4 +89,4 @@ const u8 soc_host_indexes[] = {
 	[48U] = HOST_IDX_NONE,
 	[49U] = HOST_IDX_NONE,
 };
-const u8 soc_host_indexes_sz = sizeof(soc_host_indexes);
+const u8 soc_host_indexes_sz = (u8) sizeof(soc_host_indexes);
